watcher: Rename crd package to crdwatcher
Let's use a more descriptive name. Also renamed crd.go to tracingpolicy.go.
There are no functional changes in this commit.
